requirements_m1.txt
# data science
numpy<1.20
pandas<1.4
scikit-learn<0.25
tensorflow-macos
tensorflow-metal
seaborn==0.11.2
matplotlib==3.2.2

imageio

# api
uvicorn
fastapi
starlette
Pillow
